384.4. Pulling From Multiple Contexts(Section 27: The Context System with React)
A very easy way to clean this up would be to have this button stuff right here.

The GSX get returned from a helper function so we could make a another helper function inside of your

like render button.

And maybe that would take a color as an argument.

So now I can remove the entire button GSX block into there return it and then inside the arrow function

right here.

I can return that to being just one line like so and call this render button and pass in the color.

And then I can fix up the indentation just a little bit.

So this is a slightly more clear way of doing things rather than having that kind of nasty nested function.

But still you know it can it take a little bit of time to figure out what's going on here.

Let's try saving this really quick and just making sure everything still works.

Yep everything still works if I go back over to Epcot.

J.S. I can change my value to read now and I'll see the red button appear.

All right so that's pretty much it for context at a very basic level.

We've now learned that with the context system inside of react we are sharing information remember that

is the purpose of the context system.

We have two ways of getting information into a context object and two ways of getting information out

when we are getting information out if we are consuming information from a single object.

We can set up the context property by defining that context type on our component class.

Otherwise if we want to consume multiple different context values we have to make use of the consumer

in the same way that we just saw it inside this video.

All right now we have a better idea of how context works.

Let's take a quick pause right here and continue in the next video.